+++ packages/bitcore-common/index.js
@@ -10,4 +10,4 @@
 export { default as BN } from './lib/bn.js';
 export { default as Curve } from './lib/curve/secp256k1.js';
 export * as Utils from './lib/utils.js';
-export { default as Point } from './lib/curve/short.js';
+export { default as Point } from './lib/point.js';
